requirements.txt
fastapi
uvicorn
SQLAlchemy
psycopg2-binary  # PostgreSQL adapter
python-jose[cryptography]
passlib[bcrypt]
alembic
python-dotenv
fastapi[all]
PyJWT
pydantic-settings